+++ src/test/java/nextstep/subway/path/application/PathServiceTest.java
@@ -16,8 +16,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
-import nextstep.subway.path.domain.PathFinder;
-import nextstep.subway.path.domain.ShortestPath;
 import nextstep.subway.path.dto.PathRequest;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
@@ -31,8 +29,6 @@ public class PathServiceTest {
 	LineRepository lineRepository;
 	@Mock
 	StationRepository stationRepository;
-	@Mock
-	PathFinder pathFinder;
 
 	@InjectMocks
 	PathService pathService;
@@ -74,7 +70,6 @@ public class PathServiceTest {
 		when(lineRepository.findAll()).thenReturn(Arrays.asList(신분당선, 이호선, 삼호선));
 		when(stationRepository.findById(교대역.getId())).thenReturn(Optional.of(교대역));
 		when(stationRepository.findById(양재역.getId())).thenReturn(Optional.of(양재역));
-		when(pathFinder.findShortestPath(any(), any(), any())).thenReturn(new ShortestPath(Arrays.asList(교대역, 남부터미널역, 양재역), 5L));
 
 		//when
 		PathResponse response = pathService.findShortestPath(new PathRequest(교대역.getId(), 양재역.getId()));
